+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Added support for exporting types (`export type Foo = { bar: number }`) under the `roblox` feature flag
+- Added support for using types from other modules (`local x: module.Foo`) under the `roblox` feature flag
 
 ### Fixed
 - Fixed type declaration of objects not supporting trailing commas