+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- Added support for exporting types (`export type Foo = { bar: number }`) under the `roblox` feature flag
+- Added support for using types from other modules (`local x: module.Foo`) under the `roblox` feature flag
+
+### Fixed
+- Fixed type declaration of objects not supporting trailing commas
+
 ## [0.6.2] - 2020-07-11
 ### Fixed
 - Fixed an error related with `visit_compound_op` and the `roblox` feature flag